unit tests

+++ WSEI_2022_PO_Krystian_Kuska/SQLiteAccess.cs
@@ -4,6 +4,7 @@ using System.Data;
 using System.Data.SQLite;
 using Dapper;
 using System.Linq;
+using System;
 
 namespace WSEI_2022_PO_Krystian_Kuska
 {
@@ -20,6 +21,10 @@ namespace WSEI_2022_PO_Krystian_Kuska
         //Returns score for player of id given in param
         public int LoadPlayersScores(int playerID)
         {
+            if (playerID < 1)
+            {
+                throw new ArgumentOutOfRangeException(nameof(playerID));
+            }
             using IDbConnection conn = new SQLiteConnection(LoadConnString());
             var output = conn.Query<int>($"SELECT s.Score FROM Scores s JOIN PlayerDataModel p ON s.PlayerID = p.ID WHERE s.PlayerID = {playerID};").FirstOrDefault();
             return output;
@@ -28,6 +33,14 @@ namespace WSEI_2022_PO_Krystian_Kuska
         //Checks if player of given name in param exists, if true => checks if player got new highscore, if false => inserts new player to database with their score
         public void SavePlayer(PlayerDataModel player, int score)
         {
+            if (player == null)
+            {
+                throw new ArgumentNullException(nameof(player));
+            }
+            else if (score < 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(score));
+            }
             using IDbConnection conn = new SQLiteConnection(LoadConnString());
             bool doesPlayerExist = conn.Query<int>($"SELECT COUNT(*) FROM PlayerDataModel WHERE Nickname LIKE '{player.Nickname}';").FirstOrDefault() != 0;
             if (doesPlayerExist)
@@ -45,6 +58,10 @@ namespace WSEI_2022_PO_Krystian_Kuska
         //Removes player of given nickname in param and their score
         public void DeletePlayer(string nickname)
         {
+            if (string.IsNullOrWhiteSpace(nickname) || string.IsNullOrEmpty(nickname))
+            {
+                throw new ArgumentNullException(nameof(nickname));
+            }
             using IDbConnection conn = new SQLiteConnection(LoadConnString());
             int playerID = conn.Query<int>($"SELECT ID FROM PlayerDataModel WHERE Nickname = '{nickname}'").FirstOrDefault();
             conn.Execute($"DELETE FROM Scores WHERE PlayerID = {playerID}");
